% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasek_plot.R
\name{add_liness}
\alias{add_liness}
\title{function adding bars to the strip-plot}
\usage{
add_liness(x, y, side = "both")
}
\arguments{
\item{(x, }{y) paired data}
}
\description{
function adding bars to the strip-plot
}
