% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasek_plot.R
\name{add_sds}
\alias{add_sds}
\title{function to add the whiskers left and right}
\usage{
add_sds(x, side)
}
\arguments{
\item{x}{data_vector}

\item{side}{"left" or "right"}
}
\description{
function to add the whiskers left and right
}
