% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasek_plot.R
\name{pasek_plot}
\alias{pasek_plot}
\title{this function draws a stripplot for illustrating the t or Wilcoxon paired test.}
\usage{
pasek_plot(x, y, labelx = "", labely = "", ylabel = "",
  greater = 0, color1 = ifelse(greater, "grey", "black"),
  color2 = "black", side = "both")
}
\arguments{
\item{labelx}{label on the first dataset}

\item{labely}{label on the second dataset}

\item{greater}{categorical/numeric, shows which variable is expected to be greater -- if it is zero,}

\item{color1}{the color of the less interesting case}

\item{color2}{the color of the more interesting case}

\item{(x, }{y) paired data}
}
\description{
this function draws a stripplot for illustrating the t or Wilcoxon paired test.
}
