% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasek_plot.R
\name{whistker_top_bottom}
\alias{whistker_top_bottom}
\title{function to calculate the y-coordinate of the top and bootom of the whiskers}
\usage{
whistker_top_bottom(mean_v, sd_v, topbottom)
}
\arguments{
\item{mean_v}{mean of the v vector}

\item{sd_v}{sd of the v vector}

\item{topbottom}{whether we want top or bottom}
}
\description{
function to calculate the y-coordinate of the top and bootom of the whiskers
}
